Refactoring pools comparation
