+++ feature-staking-impl/src/main/java/io/novafoundation/nova/feature_staking_impl/domain/nominationPools/selecting/SelectingNominationPoolInteractor.kt
@@ -1,19 +1,22 @@
-package io.novafoundation.nova.feature_staking_impl.domain.staking.pool
+package io.novafoundation.nova.feature_staking_impl.domain.nominationPools.selecting
 
 import io.novafoundation.nova.feature_staking_impl.data.StakingOption
+import io.novafoundation.nova.feature_staking_impl.data.chain
+import io.novafoundation.nova.feature_staking_impl.data.nominationPools.pool.KnownNovaPools
+import io.novafoundation.nova.feature_staking_impl.domain.nominationPools.common.getPoolComparator
 import io.novafoundation.nova.feature_staking_impl.domain.nominationPools.model.NominationPool
 import io.novafoundation.nova.feature_staking_impl.domain.nominationPools.model.apy
 import io.novafoundation.nova.feature_staking_impl.domain.nominationPools.pools.NominationPoolProvider
 import kotlinx.coroutines.CoroutineScope
 
 class SelectingNominationPoolInteractor(
-    private val nominationPoolProvider: NominationPoolProvider
+    private val nominationPoolProvider: NominationPoolProvider,
+    private val knownNovaPools: KnownNovaPools
 ) {
 
     suspend fun getSortedNominationPools(stakingOption: StakingOption, coroutineScope: CoroutineScope): List<NominationPool> {
         val nominationPools = nominationPoolProvider.getNominationPools(stakingOption, coroutineScope)
-        val comparator = compareByDescending<NominationPool> { it.apy }
-            .thenByDescending { it.membersCount }
+        val comparator = getPoolComparator(knownNovaPools, stakingOption.chain)
         return nominationPools.sortedWith(comparator)
     }
 }
